Load and store unknown data so it can be dealt with later.
If a type is unknown when its data is being loaded, save all data so that it can be loaded properly later.
For now, this is a one-way conversion, but it might be a good idea to also "forget" a type and unload the data back to text in the future.