Dirtiness is more visible in DarkTree now, so fix some dirtiness logic
In common:
- DefaultLeaf is dirty when created now.
- EditableObject does not contain a separate dirty flag anymore.
- DefaultEditableLeaf acts the same way now as DefaultLeaf, since it doesn't work with the EditableObject dirty flag.
- DarkTree now shows a visual indicator of dirtiness, so it's more important that we get this right.

StructureDefinition is both a leaf (the definition within the list of definitions) and a branch (the root of all the fields in the definition). We bind the dirtiness of both aspects of it to each other.

Since DefaultLeaf is dirty when created, when loading a tree of data, we just mark the root as clean once finished. Using registries to perform deferred loading may break this assumption though.